// @types/resources.d.ts
interface Resources {
// ...
  "meta": {
// ...
    "Pagrindinis": {
      "keywords": "kompiuterių remontas, naudoti kompiuteriai, verslo klasė",
      "metaDescription": "UAB Kompservisas - prekyba naudotais verslo klasės kompiuteriais. Kompiuterių remontas. Prekiaujame patikimais, bei patikrintais verlso klasės nešiojamais, bei stacionariais kompiuteriais.",
      "metaName": "UAB Kompservisas | Naudoti kompiuteriai | Algirdo g. 32",
      "name": "Pagrindinis",
      "path": "/"
    },
    "Remontas": {
      "keywords": "kompiuterių remontas, naudoti kompiuteriai, verslo klasė",
      "metaDescription": "Rinkitės mus, jei ieškote kokybės ir patirties už prieinamą kainą. Sutaisysime Jūsų naudotą nešiojamą ar stacionarų kompiuterį. Algirdo g. 32 / Šaltinių g. 2, Naujamiestis",
      "metaName": "Kompiuterių remontas Vilniuje - Kompservisas",
      "name": "Remontas",
      "path": "/remontas"
    },
// ...
  },
// ...
}